.history/requirements_20251215184451.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.10.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.3
aiofiles==23.2.1

